EZP-21213: Moved FieldStorageEvent triggering to ContentService
Since nodes are updated from ContentService::publishVersion()
*after* ContentHandler::publish() has been called, the PostPublish
FieldStorageEvent can not be called by the ContentHandler.

This commit moves it to the ContentService, and adds a sendFieldStorageEvent()
method to the ContentHandler().

It also updates the chain call by making sure the ContentHandler method
only fetches the content from the database if an event has updated it.